requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
fastapi
uvicorn
websockets
python-multipart
